requirements.txt
streamlit
torch
torchvision
transformers
pillow
accelerate
bitsandbytes